requirements.txt
aiogram==2.23.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
beautifultable==1.1.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
crypto==1.4.1
cryptography==39.0.2
cycler==0.11.0
dataframe-image==0.1.5
defusedxml==0.7.1
emoji==2.2.0
fastjsonschema==2.16.3
fonttools==4.39.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
matplotlib==3.7.1
mistune==2.0.5
multidict==6.0.4
Naked==0.1.32
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
Pillow==9.4.0
platformdirs==3.1.1
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
shellescape==3.8.1
six==1.16.0
soupsieve==2.4
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
yarl==1.8.2
